refactored

+++ Justfile
@@ -18,12 +18,13 @@ nrepl:
     clj -M:dev:nrepl
 
 dev:
-    just watch &
+    just watch 2> /dev/null &
     just plus
 
 kill:
+    #!/usr/bin/env bash
     tailwindcss-kill
-    kill `lsof -i:${PORT} -t`
+    if [[ `lsof -i:${PORT}` ]]; then kill `lsof -i:${PORT} -t`; fi
 
 run:
     just minify
